Separate accumulation of build inputs
from ContinuousBuildTriggerHandler.
