+++ subprojects/launcher/src/main/java/org/gradle/tooling/internal/provider/ContinuousBuildTriggerHandler.java
@@ -16,13 +16,6 @@
 
 package org.gradle.tooling.internal.provider;
 
-import org.gradle.api.execution.internal.TaskInputsListener;
-import org.gradle.api.internal.TaskInternal;
-import org.gradle.api.internal.file.FileCollectionInternal;
-import org.gradle.api.internal.file.FileCollectionStructureVisitor;
-import org.gradle.api.internal.file.FileTreeInternal;
-import org.gradle.api.internal.file.collections.FileSystemMirroringFileTree;
-import org.gradle.api.tasks.util.PatternSet;
 import org.gradle.execution.plan.BuildInputHierarchy;
 import org.gradle.initialization.BuildCancellationToken;
 import org.gradle.initialization.ContinuousExecutionGate;
@@ -33,21 +26,17 @@ import org.gradle.internal.os.OperatingSystem;
 import org.gradle.internal.watch.registry.FileWatcherRegistry;
 import org.gradle.internal.watch.vfs.FileChangeListener;
 
-import java.io.File;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.LinkedHashMap;
-import java.util.LinkedHashSet;
 import java.util.Map;
-import java.util.Objects;
-import java.util.Set;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
 
 import static org.gradle.internal.filewatch.DefaultFileSystemChangeWaiterFactory.QUIET_PERIOD_SYSPROP;
 
-public class ContinuousBuildTriggerHandler implements FileChangeListener, TaskInputsListener {
+public class ContinuousBuildTriggerHandler implements FileChangeListener {
     private static final boolean IS_MAC_OSX = OperatingSystem.current().isMacOsX();
 
     private final BuildInputHierarchy buildInputHierarchy;
@@ -72,10 +61,6 @@ public class ContinuousBuildTriggerHandler implements FileChangeListener, TaskIn
         this.cancellationToken = cancellationToken;
         this.continuousExecutionGate = continuousExecutionGate;
         this.quietPeriod = Long.getLong(QUIET_PERIOD_SYSPROP, 250L);
-    }
-
-    public boolean hasAnyInputs() {
-        return !buildInputHierarchy.isEmpty();
     }
 
     void wait(Runnable notifier) {
@@ -134,39 +119,6 @@ public class ContinuousBuildTriggerHandler implements FileChangeListener, TaskIn
         if (changeArrived.compareAndSet(false, true)) {
             pendingChangesListener.onPendingChanges();
         }
-    }
-
-    @Override
-    public synchronized void onExecute(TaskInternal task, FileCollectionInternal fileSystemInputs) {
-        Set<String> taskInputs = new LinkedHashSet<>();
-        Set<FilteredTree> filteredFileTreeTaskInputs = new LinkedHashSet<>();
-        fileSystemInputs.visitStructure(new FileCollectionStructureVisitor() {
-            @Override
-            public void visitCollection(FileCollectionInternal.Source source, Iterable<File> contents) {
-                contents.forEach(location -> taskInputs.add(location.getAbsolutePath()));
-            }
-
-            @Override
-            public void visitGenericFileTree(FileTreeInternal fileTree, FileSystemMirroringFileTree sourceTree) {
-                fileTree.forEach(location -> taskInputs.add(location.getAbsolutePath()));
-            }
-
-            @Override
-            public void visitFileTree(File root, PatternSet patterns, FileTreeInternal fileTree) {
-                if (patterns.isEmpty()) {
-                    taskInputs.add(root.getAbsolutePath());
-                } else {
-                    filteredFileTreeTaskInputs.add(new FilteredTree(root.getAbsolutePath(), patterns));
-                }
-            }
-
-            @Override
-            public void visitFileTreeBackedByFile(File file, FileTreeInternal fileTree, FileSystemMirroringFileTree sourceTree) {
-                taskInputs.add(file.getAbsolutePath());
-            }
-        });
-        buildInputHierarchy.recordInputs(taskInputs);
-        filteredFileTreeTaskInputs.forEach(fileTree -> buildInputHierarchy.recordFilteredInput(fileTree.getRoot(), fileTree.getPatterns().getAsSpec()));
     }
 
     private static long monotonicClockMillis() {
@@ -239,41 +191,5 @@ public class ContinuousBuildTriggerHandler implements FileChangeListener, TaskIn
         private void logOutput(StyledTextOutput logger, String message, Object... objects) {
             logger.formatln(message, objects);
         }
-    }
-
-    private static class FilteredTree {
-        private final String root;
-        private final PatternSet patterns;
-
-        private FilteredTree(String root, PatternSet patterns) {
-            this.root = root;
-            this.patterns = patterns;
-        }
-
-        public String getRoot() {
-            return root;
-        }
-
-        public PatternSet getPatterns() {
-            return patterns;
-        }
-
-        @Override
-        public boolean equals(Object o) {
-            if (this == o) {
-                return true;
-            }
-            if (o == null || getClass() != o.getClass()) {
-                return false;
-            }
-            FilteredTree that = (FilteredTree) o;
-            return root.equals(that.root) && patterns.equals(that.patterns);
-        }
-
-        @Override
-        public int hashCode() {
-            return Objects.hash(root, patterns);
-        }
-
     }
 }
